Rename setup/update scripts
This is perhaps less than ideal, but it should make the scripts work in
any web directory you put the repository into.  As such, this should be
considered part of the "initial state" trying to match Ivan's current
non-SCM setup.  We can adjust this later if he does something different
on his end.