Add A2SERVER_SCRIPT_URL to aliases
This should theoretically make the a2server-setup and a2server-update
aliases use/honor A2SERVER_SCRIPT_URL if it's set.  How this is done is
presently kind of ugly since it invokes a subshell to ensure the URL
ends in a slash, so if someone has a better idea of how to do that in a
bash alias, please file a ticket!