Fix to match Ivan's path for versionhistory.txt
This file really belongs in the A2SERVER root where people looking at
this tree can most easily find it, but for now it's more important to
follow the layout on Ivan's server.  Putting it in the root was a
mistake on my part in my initial upload because "of course".  Corrected!