Fix relative links .html -> .md
There's little reason to render these files as HTML again, and if
they're ever placed on a website, it'd probably be best to do so using a
quick bit of PHP that does the translation on the fly.  I'd suggest an
absolutely trivial stylesheet for each page (more or less akin to the
pre-HTML4 tags that were already present) and perhaps a tiny bit of
script magic that regenerates those tags based on that CSS.  Not a
priority at present.