Start of experimental install-gsos.py
This script replaces--or rather will eventually replace--the bash script
that installs netboot files in /media/A2SHARED.  It generally is going
to be implemented for now as a copy of the shell script, not quite
verbatim.  For example, rather than run individual commands with sudo,
it re-runs itself with sudo.  And rather than force scriptURL to include
a trailing slash, it explicitly chops the slash off to behave more like
you'd expect directories to.

Also, the body of installing the netboot files will be built as a
function rather than as straight script execution.  Basically the idea
there is to begin breaking the components of the netboot script into
their own functions that can be accessed independently in the future.
One step at a time there.

Anyway, very little of the script is done yet.  This is just to begin
the process and perhaps elicit some feedback if I'm really doing
something totally wrong here--my python's a little rusty.