Downloads and extracts disk 7 now
I'm not real happy with how I'm doing this.  It's obviously not as clean
or as simple as a shell script would be, but that's to be expected since
we have to emulate with those shell commands--whole programs in their
own right--do.  But right now, the logic doesn't flow right either for a
structured program or an unstructured shell script.  Probably trying to
do too much in one go there.