Different tack: Just install boot blocks
This isn't done, but it reflects largely starting over.  The goal right
now is to install the boot blocks.  To do this we need to:

* Find out if we need to install anything
* Download the disk image archive
* Extract the disk image from the archive
* Mount the image
* Copy the files out of the image
* Unmount the image
* Patch the files if they need patching
* Verify the files are patched correctly

The branching possibilities for installing the disk images are:

1. That the image is currently always a .sea.bin file.  In future, it
   might not be!  After all, the other six disk images are available
   multiple places and they're not stored as self-extracting images in
   MacBinary format.

2. That there are three different ways I know of to extract the HFS disk
   image:  Mount it as a loop filesystem (Linux), attach it as a disk
   image file (OS X), or extract its contents using hfsutils which
   should be able to run just about anywhere.  Of course hfsutils has an
   additional branch in that hfsutils "sanitizes" the files you copy out
   of it for you in ways we don't want, so we need to know the file
   names it uses so we can change them back to OS X expectations.

3. How are we installing these files?  If we're installing them for
   netatalk, we need to rename a file.  If just to a UNIX path, that's
   something different.  If there was some reason to package them in
   ProDOS format, again, we need to do something different.

At this time, since I'm developing this on a Mac, I'm assuming .sea.bin
with an 800k image, I'm implementing hdiutil to mount the image, and
I'll implement the UNIX naming first (so as Mac/Linux mount the files.)

Next I'll implement renaming for netatalk and hfsutils.  I'll check it
out on the pi and implement the Linux filesystem mounting probably last,
then clean it up to behave more like Python is supposed to if I can
before I start on the GS/OS install itself.

I need a better MacBinary unpacker than unar gives me.  The way that
works right now is just clumsy.