Begin cleanup; string formats, verbosity
This is the start of a major cleanup of the code in install-gsos.py
(which really is turning into install-a2boot.py).  First, the code is
developing three levels of output: Silent, normal, and verbose.  There
might eventually be also a fourth debug level, but that's not happening
yet.

Second, variable names just haphazardly chosen as I've been working on
this are being sanitized, made consistent, and otherwise improved.

Third (and this has only just begun), the code is being refactored a
bit for what belongs where.  That will continue.