install-gsos.py: More refactoring
I'm not real pleased with the outcome here, but all of the functions
related to mounting the HFS image and getting the a2boot files out of it
has been factored out of the bootblock installation function.

I also rewrote the disk7 download code (which has been refactored) be a
lot more robust.  It's longer than it maybe needs to be, but it's going
the right direction.