install-gsos.py: Document a useless patch
As I was preparing the patches for these boot files for 6.0.3, I
realized that p8 actually differs between GS/OS 6.0.1 and the netboot
files from Apple: The date on the splash screen is older.

This change is trivial and pedantic, but as people begin looking at
these files, how tu update them, and what to update them with, it's
worth at least noting the different version and its digest.  If someone
can make a compelling argument for the patch to be applied, we can
uncomment it.