install-gsos.py: no versioned patches, 6.0.3 P8
Originally I created the patch system in my messy Python-based monster
to allow you to patch the netboot files to the appropriate versions for
the appropriate GS/OS version, leaving 6.0.1 as pristine as possible.

But that doesn't make sense, does it?  We already fix the necessary bugs
in the AppleTalk client, and the only other patch made at this time is
the year table in ProDOS 8, and we should have that patch regardless of
what GS/OS version you install.  Of course at that point I'm one date on
the ProDOS 8 splash screen away from 6.0.3's version of P8, and I may as
well apply that to reduce the number of functionally identical files
with different SHA digests.

Long story short: Versioned patches for netboot files don't make sense,
so I removed that, and you now get 6.0.3's P8 (standalone as well as the
version embedded in "Apple //e Boot Blocks" and "ProDOS16 Image".