a2server-5-netboot.txt: Patch P8 to 6.0.3
This won't reinstall your boot blocks for you if they're not the latest
version.  If you'd like to test this (and please test this!), you'll
need to move your existing boot blocks out of the way so it can install
new ones.  Probably actually checking will happen when the Python stuff
is ready.