License terms for A2SERVER
This doesn't close the open issue on the subject, since precisely what's
under what license isn't really addressed yet, but at least the license
is specified.  For those unfamiliar with the WTFPL, it's a public domain
release like CC0, except instead of trying to legally emulate the public
domain if that concept is foreign to your jurisdiction, it basically
gives all software licenses a giant middle finger.

What it lacks in finesse and niceties or in legal rigor of the CC0, it
makes up for in simplicity: Do whatever the f*** you want with the code!
It's pretty hard to misconstrue what that means.  :)