docs/ivanx: Delete, see RasppleII/orig-archive
Having an outdated copy of the source code in a subdir of the source
code was proving to be annoying--it's bitten both Ivan and I a time or
two, and we have a GitHub organization to put repositories for these
kinds of things.  As such, this stuff is neither needed nor particularly
desirable where it is.  It lives on here:

<https://github.com/RasppleII/orig-archive/>