fix.txt: No longer necessary/desired; delete
This script well predates wheezy, and we certainly won't support
upgrades from pre-wheezy at this point.  Actually, I've made the
decision not even to support upgrades from wheezy in A2SERVER itself,
properly speaking.  We'll see if we can write an external upgrade script
(much like this one) to help make transitions a little better, but
that's the best we're going to be able to do here.