README.md: Remove section for developers (for now)
The developer information about how you can specify the source for
scripts and binaries has been having some problems, including for
upgrades.  The decision not to support upgrades from wheezy makes it
easy to stop supporting wheezy's upgrade mechanism and begin the
migration toward doing it right.  Doing it right places much greater
emphasis on upgrade security and integrity.

The developer side of this was that the scripts ran themselves straight
off a web server.  That's just a problem in general, and it too will be
fixed shortly.

The downloading of binary components is the last thing...  The right way
to do this is follow the lead of @dschmenk and package this stuff
properly which gains us much of that upgrade integrity/security I spoke
of.  That'll be a multi-step migration of its own.

This will be documented as it is implemented, but here's a heads up.