Remove a2server-update/a2server-setup aliases
The replacement update mechanism will be a script installed on your
system.  The advantage of downloading the update mechanism before
running it is that it allows you to upgrade the upgrade mechanism.
Naturally the disadvantage is that you have to download it.  Debian
systems have a mechanism for this; we'll use it.