Add worthless (but usable) install.sh
If you're installing a2server out of the source directory (eventually to
become basically the way you do it), this script is going to become the
way you do that.  For now it just runs setup/index.txt.  As we go
through things, we'll move them out of the existing scripts and into a
more modern/modular layout.