New update/index.txt
Rewrote this script—when run, upgrades will always happen from a local
tree, even if the script has to download a tree to /tmp to do it!  This
maintains an upgrade path from the a2server-update in Ivan's tree, but
allow us to phase out that kind of thing because it's insecure.

This is still effectively chipping away at the old structure, but it's a
pretty sizable chip for one new script.  Granted, that's because the old
update script didn't really do a whole lot.  *shrug*

The next chip should do similar to the one-line installation from Ivan's
website.  Once we have that, we can remove the need for scripts to
download files one at a time.  That won't remove much complexity
admittedly, but it'll remove some of it and every bit helps at this
point.